+++ tests/ValueBuilders/LookupMethodByTypeTest.php
@@ -1,0 +1,132 @@
+<?php
+
+/**
+ * Copyright (c) 2015-present Ganbaro Digital Ltd
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ *
+ *   * Redistributions of source code must retain the above copyright
+ *     notice, this list of conditions and the following disclaimer.
+ *
+ *   * Redistributions in binary form must reproduce the above copyright
+ *     notice, this list of conditions and the following disclaimer in
+ *     the documentation and/or other materials provided with the
+ *     distribution.
+ *
+ *   * Neither the names of the copyright holders nor the names of his
+ *     contributors may be used to endorse or promote products derived
+ *     from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+ * FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+ * COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+ * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+ * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+ * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+ * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ *
+ * @category  Libraries
+ * @package   Reflection/ValueBuilders
+ * @author    Stuart Herbert <stuherbert@ganbarodigital.com>
+ * @copyright 2015-present Ganbaro Digital Ltd www.ganbarodigital.com
+ * @license   http://www.opensource.org/licenses/bsd-license.php  BSD License
+ * @link      http://code.ganbarodigital.com/php-file-system
+ */
+
+namespace GanbaroDigital\Reflection\ValueBuilders;
+
+use GanbaroDigital\UnitTestHelpers\ClassesAndObjects\InvokeMethod;
+use PHPUnit_Framework_TestCase;
+
+/**
+ * @coversDefaultClass GanbaroDigital\Reflection\ValueBuilders\LookupMethodByType
+ */
+class LookupMethodByTypeTest extends PHPUnit_Framework_TestCase
+{
+    /**
+     * @covers ::lookupMethodFor
+     * @dataProvider provideDataToTest
+     */
+    public function testCanCallStatically($data, $map, $expectedResult)
+    {
+        // ----------------------------------------------------------------
+        // setup your test
+
+        // ----------------------------------------------------------------
+        // perform the change
+
+        $actualResult = InvokeMethod::onClass(LookupMethodByType_TestObject::class, 'lookupMethodFor', [$data, $map]);
+
+        // ----------------------------------------------------------------
+        // test the results
+
+        $this->assertEquals($expectedResult, $actualResult);
+    }
+
+    /**
+     * @covers ::lookupMethodFor
+     */
+    public function testReturnsFallbackWhenNothingMatches()
+    {
+        // ----------------------------------------------------------------
+        // setup your test
+
+        $data = "hello";
+        $map = [];
+        $expectedResult = "nothingMatchesTheInputType";
+
+        // ----------------------------------------------------------------
+        // perform the change
+
+        $actualResult = InvokeMethod::onClass(LookupMethodByType_TestObject::class, 'lookupMethodFor', [$data, $map]);
+
+        // ----------------------------------------------------------------
+        // test the results
+
+        $this->assertEquals($expectedResult, $actualResult);
+    }
+
+    public function provideDataToTest()
+    {
+        return $this->provideScalarTypes();
+    }
+
+    public function provideScalarTypes()
+    {
+        $dispatchMap = [
+            "Array" => "fromArray",
+            "Boolean" => "fromBoolean",
+            "Double" => "fromDouble",
+            "Integer" => "fromInteger",
+            "NULL" => "fromNull",
+            "String" => "fromString",
+        ];
+
+        return [
+            [ null, $dispatchMap, 'fromNull' ],
+            [ true , $dispatchMap, 'fromBoolean' ],
+            [ false, $dispatchMap, 'fromBoolean' ],
+            [ [ 1,2,3], $dispatchMap, 'fromArray' ],
+            [ 3.1415927, $dispatchMap, 'fromDouble' ],
+            [ 100, $dispatchMap, 'fromInteger' ],
+            [ "hello, world", $dispatchMap, 'fromString' ],
+        ];
+    }
+
+}
+
+/**
+ * gives us a container to test the trait
+ */
+class LookupMethodByType_TestObject
+{
+    use LookupMethodByType;
+}